+++ src/KernelInterpolation.jl
@@ -80,13 +80,14 @@ export NodeSet, empty_nodeset, separation_distance, dim, eachdim,
        distance_matrix, random_hypercube, random_hypercube_boundary, homogeneous_hypercube,
        homogeneous_hypercube_boundary, random_hypersphere, random_hypersphere_boundary
 export interpolation_kernel, nodeset, coefficients, kernel_coefficients,
-       polynomial_coefficients, polynomial_basis, polyvars, system_matrix,
+       polynomial_coefficients, polynomial_basis, basis, polyvars, system_matrix,
        interpolate, solve_stationary, kernel_inner_product, kernel_norm,
     kernel_matrix, operator_matrix, pde_boundary_matrix, nearest_node_index
 export rbf_fd_weights, rbf_fd_weights_at_node, rbf_fd_weights_all_nodes,
        rbf_fd_pde_matrix, rbf_fd_boundary_matrix, rbf_fd_pde_boundary_matrix,
        select_neighbors
 export Interpolation, TemporalInterpolation
+export MultiscaleInterpolation, multiscale_interpolate
 export AliveCallback, SaveSolutionCallback, SummaryCallback
 export vtk_save, vtk_read
 export examples_dir, get_examples, default_example