+++ CHANGELOG.md
@@ -9,7 +9,12 @@ for human readability.
 
 #### Added
 
-- Allow passing a factorization to `interpolate` ([#130]).
+- Add multiscale interpolation functionality with the function `multiscale_interpolate` and the type `MultiscaleInterpolation` ([#180]).
+- Allow applying differential operators to an `Interpolation` to get a callable object that evaluates the operator at any point ([#179]).
+- Added support for methods from `LinearSolve.jl` in `solve_stationary` ([#178]).
+- Added support for methods from `LinearSolve.jl` in `interpolate` ([#176]).
+- Added a keyword argument `factorization_method` to `interpolate`, `interpolation_matrix`,
+  and `least_squares_matrix` to allow for different factorization methods ([#130]).
 
 #### Changed
 
@@ -20,8 +25,6 @@ for human readability.
 
 #### Added
 
-- Added a keyword argument `factorization_method` to `interpolate`, `interpolation_matrix`,
-  and `least_squares_matrix` to allow for different factorization methods ([#130]).
 - General floating point support ([#121]).
 
 #### Changed